# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Medical Records API",
        "description": "API quan ly ho so y te",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/invoices": {
            "get": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Danh sách hóa đơn",
                "operationId": "336411b08a1358f477b049bd0ad1c6b4",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": 50
                        }
                    },
                    {
                        "name": "skip",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": 0
                        }
                    },
                    {
                        "name": "patient_id",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "number",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "start",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "payment_status",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Danh sách hóa đơn"
                    },
                    "500": {
                        "description": "Lỗi server"
                    }
                }
            },
            "post": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Tạo hóa đơn mới",
                "operationId": "600f81d4a5b7370a7fa733a8f43c6b03",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "patient_id": {
                                        "type": "string",
                                        "example": "patient_123"
                                    },
                                    "medical_record_id": {
                                        "type": "string",
                                        "example": "record_456"
                                    },
                                    "invoice_info.invoice_number": {
                                        "type": "string",
                                        "example": "INV-2024-001"
                                    },
                                    "invoice_info.invoice_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-01-15"
                                    },
                                    "payment_info.total_amount": {
                                        "type": "number",
                                        "example": 500000
                                    },
                                    "payment_status": {
                                        "type": "string",
                                        "example": "pending"
                                    },
                                    "services": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "service_type": {
                                                    "type": "string",
                                                    "example": "consultation"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Khám tổng quát"
                                                },
                                                "quantity": {
                                                    "type": "number",
                                                    "example": 1
                                                },
                                                "unit_price": {
                                                    "type": "number",
                                                    "example": 200000
                                                },
                                                "total_price": {
                                                    "type": "number",
                                                    "example": 200000
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tạo thành công"
                    },
                    "400": {
                        "description": "Dữ liệu không hợp lệ"
                    },
                    "422": {
                        "description": "Lỗi validation"
                    },
                    "500": {
                        "description": "Lỗi server"
                    }
                }
            }
        },
        "/api/v1/invoices/{id}": {
            "get": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Chi tiết hóa đơn",
                "operationId": "023f984f1e989be310a4aa4e47d3fff5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thông tin hóa đơn"
                    },
                    "404": {
                        "description": "Không tìm thấy"
                    },
                    "500": {
                        "description": "Lỗi server"
                    }
                }
            },
            "put": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Cập nhật hóa đơn",
                "operationId": "dfcabd73d87b50d44c217a1148f2062a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "_rev": {
                                        "type": "string",
                                        "example": "1-abc123"
                                    },
                                    "payment_status": {
                                        "type": "string",
                                        "example": "paid"
                                    },
                                    "paid_date": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cập nhật thành công"
                    },
                    "400": {
                        "description": "Dữ liệu không hợp lệ"
                    },
                    "404": {
                        "description": "Không tìm thấy"
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "500": {
                        "description": "Lỗi server"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Xóa hóa đơn",
                "operationId": "b6ceb60ff2fb3d4207bcd2b377ea87b4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "rev",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xóa thành công"
                    },
                    "400": {
                        "description": "Thiếu rev parameter"
                    },
                    "404": {
                        "description": "Không tìm thấy"
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "500": {
                        "description": "Lỗi server"
                    }
                }
            }
        },
...
    },
...
}